+++ assembler/src/include/omni/omni_utils.hpp
@@ -1998,7 +1998,7 @@ inline size_t PairInfoPathLengthUpperBound(size_t k, size_t insert_size,
 }
 
 inline size_t PairInfoPathLengthLowerBound(size_t k, size_t l_e1, size_t l_e2,
-		size_t gap, double delta) {
+		int gap, double delta) {
 	double answer = 0. + gap + k + 2 - l_e1 - l_e2 - delta;
 	return math::gr(answer, 0.) ? std::floor(answer) : 0;
 }
